move all files and fix testing
