+++ atomate2/openmm/jobs/nvt_maker.py
@@ -1,8 +1,7 @@
-from src.atomate2.openmm.jobs.base_openmm_maker import BaseOpenMMMaker
-from src.atomate2.openmm.schemas.task_details import TaskDetails
-from openmm.openmm import AndersenThermostat
+from atomate2.openmm.jobs.base_openmm_maker import BaseOpenMMMaker
+from atomate2.openmm.schemas.task_details import TaskDetails
 from dataclasses import dataclass
-from openmm.unit import kelvin, pico, second
+from openmm.unit import kelvin
 
 
 @dataclass
