+++ atomate2/openmm/schemas/physical_state.py
