+++ atomate2/openmm/schemas/task_details.py
