+++ tests/openmm/flows/test_anneal_maker.py
@@ -1,13 +1,10 @@
-from src.atomate2.openmm.flows.anneal_maker import AnnealMaker
-from src.atomate2.openmm.jobs.temp_change_maker import TempChangeMaker
+from atomate2.openmm.flows.anneal_maker import AnnealMaker
+from atomate2.openmm.jobs.temp_change_maker import TempChangeMaker
 
 
 def test_anneal_maker(alchemy_input_set, job_store):
-    from src.atomate2.openmm.jobs.energy_minimization_maker import EnergyMinimizationMaker
-    from src.atomate2.openmm.jobs.nvt_maker import NVTMaker
-    from src.atomate2.openmm.jobs.npt_maker import NPTMaker
-    from src.atomate2.openmm.flows.production_maker import ProductionMaker
-    from src.atomate2.openmm.schemas.openmm_task_document import OpenMMTaskDocument
+    from atomate2.openmm.jobs.nvt_maker import NVTMaker
+    from atomate2.openmm.schemas.openmm_task_document import OpenMMTaskDocument
     from jobflow import run_locally
 
     anneal_maker = AnnealMaker(
