+++ tests/openmm/schemas/test_task_document.py
@@ -1,3 +1,2 @@
 def test_openmm_task_document():
-    from src.atomate2.openmm.schemas.openmm_task_document import OpenMMTaskDocument
     pass
